Create Data Acquisition Requirements
Perform the following tasks to combine an existing data set with some external data:



A.  Summarize a research question that can be answered using both the original database and the add-on CSV data. The question should require data from both these data sources.

1. Identify which data from the original data set and the add-on CSV file are needed to answer the research question.


B.  Create a logical data model for the add-on CSV file by evaluating the data contained in the file and emphasizing the relational constraints.

1. Write SQL code that creates a table that accommodates the extension of the logical data model to a physical data model by specifying the field types and relevant keys.

2. Write SQL code that loads the data from the add-on CSV file into the table created in part B1.


C.  Write SQL statement(s) for a query or queries that inform the research question summarized in part A.

1. Provide a CSV file or files that capture the results from the query or queries.


D.  Determine how often the add-on file should be acquired and refreshed in the database for the data to remain relevant to the business and the research question.


E.  Create an SQL script that performs the process of loading the add-on data.


F.  Provide a Panopto video recording that includes a demonstration of the functionality of the code used for the analysis and a summary of the programming environment.


Note: For instructions on how to access and use Panopto, use the "Panopto How-To Videos" web link provided below. To access Panopto's website, navigate to the web link titled "Panopto Access" and then choose to log in using the “WGU” option. If prompted, log in using your WGU student portal credentials, and then it will forward you to Panopto’s website.


To submit your recording, upload it to the Panopto drop box titled “Master of Science, Data Analytics TGM1 | D205 (Student Creators) [assignments]." Once the recording has been uploaded and processed in Panopto's system, retrieve the URL of the recording from Panopto and copy and paste it into the Links option. Upload the remaining task requirements using the Attachments option.


G.  Record the web sources used to acquire data or segments of third-party code to support the application. Be sure the web sources are reliable.


H.  Acknowledge sources, using in-text citations and references, for content that is quoted, paraphrased, or summarized.


I.  Demonstrate professional communication in the content and presentation of your submission.